cleaned project files with remove. removed cqrsrepository from ef and cosmosdb. tried cosmosdb as the primary db
