--- samples/BoltOn.Samples.Application/Entities/Grade.cs
@@ -1,17 +1,0 @@
-﻿using BoltOn.Data;
-using Newtonsoft.Json;
-
-namespace BoltOn.Samples.Application.Entities
-{
-    public class Grade : BaseEntity<string>
-    {
-        //while using any property in cosmosdb query make sure propertyname matches exactly as it is in stored in document collection.
-        [JsonProperty("id")]
-        public override string Id { get; set; }
-        [JsonProperty("studentId")]
-        public int StudentId { get; set; }
-        public string CourseName { get; set; }
-        public int Year { get; set; }
-        public string Score { get; set; }
-    }
-}
