+++ samples/BoltOn.Samples.Infrastructure/BoltOn.Samples.Infrastructure.csproj
@@ -11,9 +11,6 @@
         <NoStdLib>true</NoStdLib>
     </PropertyGroup>
     <ItemGroup>
-        <Compile Remove="Class1.cs" />
-    </ItemGroup>
-    <ItemGroup>
         <Folder Include="Data\" />
         <Folder Include="Data\Mappings\" />
         <Folder Include="Data\Repositories\" />
