--- src/BoltOn.Data.CosmosDb/BaseCosmosDbContext.cs
@@ -1,12 +1,0 @@
-﻿namespace BoltOn.Data.CosmosDb
-{
-	public abstract class BaseCosmosDbContext<TCosmosDbContext> where TCosmosDbContext : BaseCosmosDbContext<TCosmosDbContext>
-    {
-		public CosmosDbContextOptions<TCosmosDbContext> Options { get; private set; }
-
-		protected BaseCosmosDbContext(CosmosDbContextOptions<TCosmosDbContext> options)
-        {
-			Options = options;
-        }
-    }
-}
