+++ src/BoltOn.Data.CosmosDb/BoltOn.Data.CosmosDb.csproj
@@ -24,10 +24,4 @@
     <ItemGroup>
         <ProjectReference Include="..\BoltOn\BoltOn.csproj" />
     </ItemGroup>
-    <ItemGroup>
-        <Compile Remove="CosmosDbSettings.cs" />
-        <Compile Remove="BaseCosmosDbContext.cs" />
-        <Compile Remove="BaseCosmosDbRepository.cs" />
-        <Compile Remove="CosmosDbContextOptions.cs" />
-    </ItemGroup>
 </Project>
