--- src/BoltOn/BoltOn.nuspec
@@ -1,25 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2013/05/nuspec.xsd">
-  <metadata>
-    <id>BoltOn</id>
-    <version>0.3.0</version>
-    <title>BoltOn</title>
-    <authors>Gokulnath Murugesan</authors>
-    <owners>Gokulnath Murugesan</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <projectUrl>https://github.com/gokulm/BoltOn</projectUrl>
-    <description>BoltOn is an open source framework to build any .NET application with proper separation of concerns.</description>
-    <summary></summary>
-    <releaseNotes></releaseNotes>
-    <copyright></copyright>
-    <language></language>
-    <tags>C#, .NET Core</tags>
-    <dependencies>
-      <group targetFramework=".NETStandard2.0">
-        <dependency id="Microsoft.CSharp" version="4.5.0" />
-        <dependency id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.2.0" />
-        <dependency id="Microsoft.Extensions.Logging.Abstractions" version="2.2.0" />
-      </group>
-    </dependencies>
-  </metadata>
-</package>
