+++ tests/BoltOn.Tests/Other/StudentRepository.cs
@@ -1,5 +1,7 @@
-﻿using BoltOn.Data;
+﻿using BoltOn.Cqrs;
+using BoltOn.Data;
 using BoltOn.Data.EF;
+using BoltOn.Utilities;
 
 namespace BoltOn.Tests.Other
 {
@@ -9,7 +11,8 @@ namespace BoltOn.Tests.Other
 
 	public class StudentRepository : Repository<Student, SchoolDbContext>, IStudentRepository
     {
-        public StudentRepository(IDbContextFactory dbContextFactory) : base(dbContextFactory)
+        public StudentRepository(IDbContextFactory dbContextFactory, EventBag eventBag,
+			IBoltOnClock boltOnClock) : base(dbContextFactory, eventBag, boltOnClock)
         {
         }
     }
